+++ main.js
@@ -230,6 +230,7 @@ function initializeGame() {
         console.log('[v0.47.0 B1] currentMap restored:', gameState.currentMap ? gameState.currentMap.name : 'null');
     }
 
+    if (typeof _initXpPool === 'function') _initXpPool();
     gameState.spawnProtectUntil    = 0;
     if (typeof resetPackNames === 'function') resetPackNames();
     gameState.mutationPanelOpen    = false;